refactor: 테스트 패키지의 도메인 코드를 테스트 코드와 분리
