+++ src/main/kotlin/study/Language.kt
@@ -1,0 +1,3 @@
+package study
+
+data class Language(val type: String, val level: Int)
