+++ src/main/kotlin/study/Languages.kt
@@ -1,0 +1,3 @@
+package study
+
+class Languages(val value: List<Language>)
