+++ src/main/kotlin/study/PersonBuilder.kt
@@ -1,0 +1,32 @@
+package study
+
+class PersonBuilder {
+    private lateinit var name: String
+    private var company: String? = null
+    private var skills: Skills? = null
+    private var languages: Languages? = null
+
+    fun name(value: String) {
+        this.name = value
+    }
+
+    fun company(value: String) {
+        this.company = value
+    }
+
+    fun skills(block: SkillsBuilder.() -> Unit) {
+        this.skills = SkillsBuilder().apply(block).build()
+    }
+
+    fun languages(block: LanguagesBuilder.() -> Unit) {
+        this.languages = LanguagesBuilder().apply(block).build()
+    }
+
+    fun build(): Person {
+        return Person(name, company, skills, languages)
+    }
+}
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    return PersonBuilder().apply(block).build()
+}
